Move task group index slicing to TaskInstance
This logic will be reused by TriggerRuleDep.
